backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=1.5.0
scikit-learn>=1.0.0
xgboost>=1.6.0
joblib>=1.2.0
beautifulsoup4>=4.11.0
requests>=2.28.0
python-whois>=0.8.0
googlesearch-python>=1.0.0
urllib3>=1.26.0
lxml>=4.9.0
python-dateutil>=2.8.0